fix(a2): R3 runtime training-source fixes for formal training
- door_open_a2_base.py: reward_scales config fallback (__init__ ordering), reward-coverage expected follows raw_components, termination dtype relax + scaled bool->float, evidence step_index clamp+step_mask for reset boundary, valid_hinge excludes reset-boundary envs
- legged_robot_base.py: _reward_termination returns .float()
- train_agent_trl.py: unwrap ACTIVE_SOURCE_LOCK wrapper to validate inner SOURCE_FROZEN lock
- average_meters.py: bool tensor -> float before mean

+++ gr00t/rl/envs/legged_base_task/legged_robot_base.py
@@ -1992,7 +1992,7 @@ class LeggedRobotBase(BaseTask):
 
     def _reward_termination(self):
         # Terminal reward / penalty
-        return self.reset_buf * ~self.time_out_buf
+        return (self.reset_buf * ~self.time_out_buf).float()
 
     def _reward_penalty_torques(self):
         # Penalize torques
